chore: reorganize packages
