+++ cmd/schema_new.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/cueblox/blox/schema"
+	"github.com/cueblox/blox/internal/schema"
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
 )
