+++ internal/hosting/netlify/package.json
